chore: moved notification components to different layers
