--- src/renderer/entities/notification/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './NotificationRow/NotificationRow';
