+++ src/renderer/features/notifications/index.ts
@@ -1,0 +1,1 @@
+export { NotificationProvider } from './NotificationProvider/NotificationProvider';
