+++ src/renderer/pages/Notifications/Notifications.tsx
@@ -5,8 +5,9 @@ import { useI18n } from '@app/providers';
 import EmptyNotifications from './components/EmptyNotifications';
 import { sortByDate } from './common/utils';
 import { FootnoteText } from '@shared/ui';
-import { useNotification, NotificationRow } from '@entities/notification';
+import { useNotification } from '@entities/notification';
 import { Header } from '@renderer/components/common';
+import { NotificationRow } from './components/NotificationRow';
 
 export const Notifications = () => {
   const { t, dateLocale } = useI18n();
